Compile with GCC 13
The E4S configuration currently sets the compiler to GCC 11.
However, the standard C++ library is not compatible with the
code compiled by the clang compiler in ROCm 6.2. I needed to
bump up the GCC compiler to version 13. This required adding
a new compiler definition. Ideally, this would be added to
the general compilers.yaml. However, I wanted to avoid making
change directly to the E4S configuration, so I added the
changes to the configuration for the ROCm compiler.