Reenable installing luaposix
The spack paraview package is for some reason loading the lua
package when it loads paraview. However, the `srun` command
also uses lua, and it needs the luaposix package that is not
installed with spack by default.

I had fixed this problems a while ago. I then removed the fix
because the paraview load no longer seemed to be loading lua.
But now it is apparently loading lua again, so put the fix
back in.