Fix issue with automake
Frontier has automake 1.15 installed, but the autotools being installed
by Spack needs 1.16. For some packages being built with autotools, they
do not properly depend on automake and cannot find the appropriate
version of aclocal. The fix was adding automake to the environment. (I'm
not sure how Spack figures out that should get installed everywhere.)
I also added configuration to make packages depend on the correct
version of automake. This is probably not necessary, but it is the first
thing I tried and together they work.